src/modules/users/services/AuthenticateUserService.spec.ts
// import AppError from '@shared/errors/AppError';
import FakeUsersRepository from '../repositories/fakes/FakeUsersRepository';
import FakeHashProvider from '../providers/hashProvider/fakes/FakeHashProvider';
import AuthenticateUserService from './AuthenticateUserService';
import CreateUserService from './CreateUserService';

describe('AuthenticateUser', () => {
  it('should be able to authenticate', async () => {
    const fakeUsersRepository = new FakeUsersRepository();
    const fakeHashProvider = new FakeHashProvider();

    const createUserService = new CreateUserService(
      fakeUsersRepository,
      fakeHashProvider,
    );

    const authenticateUserService = new AuthenticateUserService(
      fakeUsersRepository,
      fakeHashProvider,
    );
    // creating the user first and encrypting the password
    const user = await createUserService.execute({
      name: 'John Doe',
      email: 'johndoe@example.com',
      password: '123456',
    });

    const response = await authenticateUserService.execute({
      email: 'johndoe@example.com',
      password: '123456',
    });

    expect(response).toHaveProperty('token');
    expect(response.user).toEqual(user);
  });

  // it('should not be able to create a new user with same email from another user', async () => {
  //   const fakeUsersRepository = new FakeUsersRepository();
  //   const authenticateUserService = new AuthenticateUserService(
  //     fakeUsersRepository,
  //   );

  //   const email = 'johndoe@example.com';

  //   await authenticateUserService.execute({
  //     name: 'John Doe',
  //     email,
  //     password: '123456',
  //   });

  //   // creating a second user with the same email of the first one
  //   expect(
  //     createUserService.execute({
  //       name: 'John Doe',
  //       email,
  //       password: '123456',
  //     }),
  //   ).rejects.toBeInstanceOf(AppError);
  // });
});
